Project modernization

+++ src/Compilers/Core/Portable/Collections/ImmutableSegmentedDictionary`2.cs
@@ -20,7 +20,7 @@ namespace Loretta.CodeAnalysis.Collections
     ///
     /// <para>The following table summarizes the performance characteristics of
     /// <see cref="ImmutableSegmentedDictionary{TKey, TValue}"/>:</para>
-    /// 
+    ///
     /// <list type="table">
     ///   <item>
     ///     <description>Operation</description>
@@ -41,7 +41,7 @@ namespace Loretta.CodeAnalysis.Collections
     ///     <description>Requires creating a new segmented dictionary</description>
     ///   </item>
     /// </list>
-    /// 
+    ///
     /// <para>This type is backed by segmented arrays to avoid using the Large Object Heap without impacting algorithmic
     /// complexity.</para>
     /// </remarks>
@@ -70,9 +70,9 @@ namespace Loretta.CodeAnalysis.Collections
 
         private readonly SegmentedDictionary<TKey, TValue> _dictionary;
 
-        private ImmutableSegmentedDictionary(SegmentedDictionary<TKey, TValue> dictionary!!)
+        private ImmutableSegmentedDictionary(SegmentedDictionary<TKey, TValue> dictionary)
         {
-            _dictionary = dictionary;
+            _dictionary = dictionary ?? throw new ArgumentNullException(nameof(dictionary));
         }
 
         public IEqualityComparer<TKey> KeyComparer => _dictionary.Comparer;
@@ -214,8 +214,9 @@ namespace Loretta.CodeAnalysis.Collections
             return new ImmutableSegmentedDictionary<TKey, TValue>(dictionary);
         }
 
-        public ImmutableSegmentedDictionary<TKey, TValue> RemoveRange(IEnumerable<TKey> keys!!)
+        public ImmutableSegmentedDictionary<TKey, TValue> RemoveRange(IEnumerable<TKey> keys)
         {
+            if (keys is null) throw new ArgumentNullException(nameof(keys));
             var result = ToBuilder();
             result.RemoveRange(keys);
             return result.ToImmutable();
@@ -236,8 +237,9 @@ namespace Loretta.CodeAnalysis.Collections
             return new ImmutableSegmentedDictionary<TKey, TValue>(dictionary);
         }
 
-        public ImmutableSegmentedDictionary<TKey, TValue> SetItems(IEnumerable<KeyValuePair<TKey, TValue>> items!!)
+        public ImmutableSegmentedDictionary<TKey, TValue> SetItems(IEnumerable<KeyValuePair<TKey, TValue>> items)
         {
+            if (items is null) throw new ArgumentNullException(nameof(items));
             var result = ToBuilder();
             foreach (var item in items)
             {
